Move from Pootle to Zing in most references
This commit adds the documentation for `zing` TS plugin,
References this plugin as a default one in the sample config,
and switches to mentioning Zing in most of the documentation.

Pootle plugin documentation page and Zing + Pootle guide
are left for historic purposes (to avoid broken links) with
an appropriate deprecation warning.